+++ build_for_selfware.sh
@@ -7,6 +7,12 @@ set -e
 npm run build:es
 npm run build:css
 
-./node_modules/.bin/rimraf ../Selfware/selfware/node_modules/react-infinite-table/dist/
+SELFWARE_PATH=../Selfware/selfware
 
-cp -r dist ../Selfware/selfware/node_modules/react-infinite-table/dist
+./node_modules/.bin/rimraf "$SELFWARE_PATH/node_modules/react-infinite-table/dist/"
+./node_modules/.bin/rimraf "$SELFWARE_PATH/node_modules/react-infinite-table/src/"
+
+cp -r dist "$SELFWARE_PATH/node_modules/react-infinite-table/dist"
+
+mkdir "$SELFWARE_PATH/node_modules/react-infinite-table/src"
+cp -r src/style.scss "$SELFWARE_PATH/node_modules/react-infinite-table/src/"