+++ src/lib/Cell.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-export default class Cell extends React.PureComponent {
+export default class Cell extends React.Component {
   static propTypes = {
     renderer: PropTypes.func.isRequired,
     columnIndex: PropTypes.number,