+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+##### 1.3.0
+
+- Remove `key` from renderer callbacks, the `key` is handled internally
+- Changed the way React `key` is computed for `<tr>`s:
+  - if the row was render before, use the same key
+  - otherwise, try to reuse an existing row that was removed from the visible window
+
 ##### 1.2.3
 
 - Added prop: use `fillTableWidth = true` to grow columns to fill the width of the table